refactor: move helper to saved search folder
