+++ src/lib/Components/Artist/ArtistArtworks/SavedSearchBanner.tsx
@@ -4,7 +4,8 @@ import { SavedSearchBanner_me } from "__generated__/SavedSearchBanner_me.graphql
 import { SavedSearchBannerCreateSavedSearchMutation } from "__generated__/SavedSearchBannerCreateSavedSearchMutation.graphql"
 import { SavedSearchBannerDisableSavedSearchMutation } from "__generated__/SavedSearchBannerDisableSavedSearchMutation.graphql"
 import { SavedSearchBannerQuery } from "__generated__/SavedSearchBannerQuery.graphql"
-import { FilterParams, prepareFilterParamsForSaveSearchInput } from "lib/Components/ArtworkFilter/ArtworkFilterHelpers"
+import { FilterParams } from "lib/Components/ArtworkFilter/ArtworkFilterHelpers"
+import { prepareFilterParamsForSaveSearchInput } from 'lib/Components/ArtworkFilter/SavedSearch/searchCriteriaHelpers'
 import { SearchCriteriaAttributes } from "lib/Components/ArtworkFilter/SavedSearch/types"
 import { usePopoverMessage } from "lib/Components/PopoverMessage/popoverMessageHooks"
 import { LegacyNativeModules } from "lib/NativeModules/LegacyNativeModules"
