+++ src/lib/Components/ArtworkFilter/SavedSearch/constants.ts
@@ -1,0 +1,18 @@
+export const allowedSearchCriteriaKeys = [
+  "artistID",
+  "locationCities",
+  "colors",
+  "partnerIDs",
+  "additionalGeneIDs",
+  "attributionClass",
+  "majorPeriods",
+  "acquireable",
+  "atAuction",
+  "inquireableOnly",
+  "offerable",
+  "materialsTerms",
+  "priceRange",
+  "dimensionRange",
+  "height",
+  "width",
+]
